[fix] CRA-82 관리자 자신 삭제 예외 처리

+++ src/main/java/com/yoyomo/domain/club/domain/service/ClubManagerDeleteService.java
@@ -1,5 +1,6 @@
 package com.yoyomo.domain.club.domain.service;
 
+import com.yoyomo.domain.application.exception.ManagerDeleteException;
 import com.yoyomo.domain.club.domain.repository.ClubMangerRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,7 +13,10 @@ public class ClubManagerDeleteService {
 
     private final ClubMangerRepository clubMangerRepository;
 
-    public void delete(List<Long> clubManagerIds) {
+    public void delete(List<Long> clubManagerIds, long userId) {
+        if (clubManagerIds.contains(userId)) {
+            throw new ManagerDeleteException();
+        }
         clubMangerRepository.deleteAllById(clubManagerIds);
     }
 }
